+++ zetaclient/common/constant.go
@@ -14,4 +14,7 @@ const (
 
 	// RPCStatusCheckInterval is the interval to check RPC status, 1 minute
 	RPCStatusCheckInterval = time.Minute
+
+	// BTCMempoolStuckTxCheckInterval is the interval to check for Bitcoin stuck transactions in the mempool
+	BTCMempoolStuckTxCheckInterval = 30 * time.Second
 )